requirements.txt
pandas
polygon-api-client>=1.12.0
# polygon-python-client